requirements.txt
bcrypt==4.2.1

pyjwt==2.10.1

cryptography==44.0.1